requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
langchain==0.1.16
langchain-community==0.0.34
Pillow==10.3.0
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.20.0
requests==2.31.0
langchain-experimental==0.0.57
scikit-learn==1.4.0
